extract MockPlayer into separate file
